requirements.txt
gTTS
FFMPEG
ollama
PyAudio
pydub
pygame
python-dotenv
SpeechRecognition